Movement + Static Knockback script
Movement script is referenced. You just drag and drop the movement script onto the player, then drag and drop one of the preset scriptable objects. "Player Movement Data" is my preferred one.

Knockback is a public static script;
Just call KnockBack.Begin
(Rigidbody, direction, forceamount)
Pass it the Rigidbody2d being affected
Direction -1 for left knockback
                 1 for right
forceamount is how much force you want to apply, it needs to be 15+ on the player to be significant but because it is a Impulse it will change depending on the objects mass.